meterUpload/requirements.txt
asn1crypto==0.24.0
Babel==2.7.0
blessings==1.7
blinker==1.4
bpython==0.18
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
coverage==4.5.4
crcmod==1.7
cryptography==2.7
curtsies==0.3.0
decorator==4.4.0
defusedxml==0.6.0
Django==1.10.3
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.0
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
flipflop==1.0
greenlet==0.4.15
guess-language==0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.3
pandas==0.24.2
pbr==5.4.2
pep8==1.7.1
pipenv==2018.11.26
pycparser==2.19
pycryptodome==3.8.2
Pygments==2.4.2
PyMySQL==0.9.3
pyserial==3.4
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
redis==3.3.6
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.8
sqlalchemy-migrate==0.12.0
sqlparse==0.3.0
Tempita==0.5.2
typing==3.7.4
urllib3==1.25.3
virtualenv==16.6.2
virtualenv-clone==0.5.3
wcwidth==0.1.7
Werkzeug==0.15.5
Whoosh==2.7.4
WTForms==2.2.1
xlrd==1.2.0
